Refactor more util functions
