+++ packages/protocol_cosmwasm/src/utils.rs
@@ -1,5 +1,12 @@
 use cosmwasm_std::{StdError, Uint128};
 
+// Slice the length of the bytes array into 32bytes
+pub fn element_encoder(v: &[u8]) -> [u8; 32] {
+    let mut output = [0u8; 32];
+    output.iter_mut().zip(v).for_each(|(b1, b2)| *b1 = *b2);
+    output
+}
+
 // Computes the combination bytes of "chain_type" and "chain_id".
 // Combination rule: 8 bytes array(00 * 2 bytes + [chain_type] 2 bytes + [chain_id] 4 bytes)
 // Example:
@@ -10,7 +17,7 @@ pub fn compute_chain_id_type(chain_id: u64, chain_type: &[u8]) -> u64 {
     let mut buf = [0u8; 8];
     #[allow(clippy::needless_borrow)]
     buf[2..4].copy_from_slice(&chain_type);
-    buf[4..8].copy_from_slice(&chain_id_value.to_be_bytes());
+    buf[4..8].copy_from_slice(&chain_id_value.to_le_bytes());
     u64::from_be_bytes(buf)
 }
 
@@ -27,4 +34,12 @@ pub fn parse_string_to_uint128(v: String) -> Result<Uint128, StdError> {
         Err(e) => return Err(StdError::GenericErr { msg: e.to_string() }),
     };
     Ok(res)
+}
+
+// Get the `chain_id_type` from bytes array.
+pub fn get_chain_id_type(chain_id_type: &[u8]) -> u64 {
+    let mut buf = [0u8; 8];
+    #[allow(clippy::needless_borrow)]
+    buf[2..8].copy_from_slice(&chain_id_type);
+    u64::from_be_bytes(buf)
 }
